requirements.txt
streamlit
tensorflow>=2.11
efficientnet
opencv-python
Pillow
numpy
pandas
scikit-learn
matplotlib
fpdf
tqdm
